Can disallow attributes
And even re-allow them back with a config similar to re-allowing method calls.

Close #179
