+++ src/DisallowedAttributeFactory.php
@@ -3,13 +3,26 @@ declare(strict_types = 1);
 
 namespace Spaze\PHPStan\Rules\Disallowed;
 
+use Spaze\PHPStan\Rules\Disallowed\Exceptions\UnsupportedParamTypeInConfigException;
+
 class DisallowedAttributeFactory
 {
+
+	/** @var Allowed */
+	private $allowed;
+
+
+	public function __construct(Allowed $allowed)
+	{
+		$this->allowed = $allowed;
+	}
+
 
 	/**
 	 * @param array $config
 	 * @phpstan-param DisallowedAttributesConfig $config
 	 * @return DisallowedAttribute[]
+	 * @throws UnsupportedParamTypeInConfigException
 	 */
 	public function createFromConfig(array $config): array
 	{
@@ -22,6 +35,7 @@ class DisallowedAttributeFactory
 					$disallowed['message'] ?? null,
 					$disallowed['allowIn'] ?? [],
 					$disallowed['allowExceptIn'] ?? $disallowed['disallowIn'] ?? [],
+					$this->allowed->getConfig($disallowed),
 					$disallowed['errorIdentifier'] ?? null,
 					$disallowed['errorTip'] ?? null
 				);
