+++ tests/Calls/FunctionCallsNamedParamsTest.php
@@ -11,6 +11,7 @@ use Spaze\PHPStan\Rules\Disallowed\Allowed;
 use Spaze\PHPStan\Rules\Disallowed\AllowedPath;
 use Spaze\PHPStan\Rules\Disallowed\DisallowedCallFactory;
 use Spaze\PHPStan\Rules\Disallowed\Formatter\Formatter;
+use Spaze\PHPStan\Rules\Disallowed\Normalizer\Normalizer;
 use Spaze\PHPStan\Rules\Disallowed\RuleErrors\DisallowedRuleErrors;
 
 /**
@@ -25,9 +26,11 @@ class FunctionCallsNamedParamsTest extends RuleTestCase
 	protected function getRule(): Rule
 	{
 		$formatter = new Formatter();
+		$normalizer = new Normalizer();
+		$allowed = new Allowed($formatter, $normalizer, new AllowedPath(new FileHelper(__DIR__)));
 		return new FunctionCalls(
-			new DisallowedRuleErrors(new Allowed($formatter, new AllowedPath(new FileHelper(__DIR__)))),
-			new DisallowedCallFactory($formatter),
+			new DisallowedRuleErrors($allowed),
+			new DisallowedCallFactory($formatter, $normalizer, $allowed),
 			[
 				[
 					'function' => 'Foo\Bar\Waldo\foo()',
