+++ tests/Calls/FunctionCallsUnsupportedParamConfigTest.php
@@ -10,6 +10,7 @@ use Spaze\PHPStan\Rules\Disallowed\Allowed;
 use Spaze\PHPStan\Rules\Disallowed\AllowedPath;
 use Spaze\PHPStan\Rules\Disallowed\DisallowedCallFactory;
 use Spaze\PHPStan\Rules\Disallowed\Formatter\Formatter;
+use Spaze\PHPStan\Rules\Disallowed\Normalizer\Normalizer;
 use Spaze\PHPStan\Rules\Disallowed\RuleErrors\DisallowedRuleErrors;
 
 class FunctionCallsUnsupportedParamConfigTest extends PHPStanTestCase
@@ -23,9 +24,11 @@ class FunctionCallsUnsupportedParamConfigTest extends PHPStanTestCase
 		$this->expectException(ShouldNotHappenException::class);
 		$this->expectExceptionMessage('{foo(),bar()}: Parameter #2 $definitelyNotScalar has an unsupported type array specified in configuration');
 		$formatter = new Formatter();
+		$normalizer = new Normalizer();
+		$allowed = new Allowed($formatter, $normalizer, new AllowedPath(new FileHelper(__DIR__)));
 		new FunctionCalls(
-			new DisallowedRuleErrors(new Allowed($formatter, new AllowedPath(new FileHelper(__DIR__)))),
-			new DisallowedCallFactory($formatter),
+			new DisallowedRuleErrors($allowed),
+			new DisallowedCallFactory($formatter, $normalizer, $allowed),
 			[
 				[
 					'function' => [
