+++ tests/Configs/InsecureConfigMethodCallsTest.php
@@ -13,6 +13,7 @@ use Spaze\PHPStan\Rules\Disallowed\AllowedPath;
 use Spaze\PHPStan\Rules\Disallowed\Calls\MethodCalls;
 use Spaze\PHPStan\Rules\Disallowed\DisallowedCallFactory;
 use Spaze\PHPStan\Rules\Disallowed\Formatter\Formatter;
+use Spaze\PHPStan\Rules\Disallowed\Normalizer\Normalizer;
 use Spaze\PHPStan\Rules\Disallowed\RuleErrors\DisallowedMethodRuleErrors;
 use Spaze\PHPStan\Rules\Disallowed\RuleErrors\DisallowedRuleErrors;
 use Spaze\PHPStan\Rules\Disallowed\Type\TypeResolver;
@@ -28,13 +29,15 @@ class InsecureConfigMethodCallsTest extends RuleTestCase
 		// Load the configuration from this file
 		$config = Neon::decode(file_get_contents(__DIR__ . '/../../disallowed-insecure-calls.neon'));
 		$formatter = new Formatter();
+		$normalizer = new Normalizer();
+		$allowed = new Allowed($formatter, $normalizer, new AllowedPath(new FileHelper(__DIR__)));
 		return new MethodCalls(
 			new DisallowedMethodRuleErrors(
-				new DisallowedRuleErrors(new Allowed($formatter, new AllowedPath(new FileHelper(__DIR__)))),
+				new DisallowedRuleErrors($allowed),
 				new TypeResolver(),
 				$formatter
 			),
-			new DisallowedCallFactory($formatter),
+			new DisallowedCallFactory($formatter, $normalizer, $allowed),
 			$config['parameters']['disallowedMethodCalls']
 		);
 	}
