+++ tests/libs/AttributeEntity.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace Foo\Bar;
+namespace Attributes;
 
 use Attribute;
